% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_boot_sum.R
\name{ci_boot_sum}
\alias{ci_boot_sum}
\title{Compute summed bootstrapped confidence intervals}
\arguments{
\item{dat}{Dataset of juvenile or adult salmon}

\item{by}{Break dat by week, month or year}

\item{var}{Numeric variable to compute to estimate sum and confidence intervals}

\item{conf.int}{Confidence intervals such as 68, 80, 90, 95, 99}
}
\value{
Summed fish passage and confidence intervals on the sum
}
\description{
Compute summed bootstrapped confidence intervals
}
\examples{
library(BD)
library(tidyverse)

data(fish)
salmon <- filter(fish, wk < 6)
salmon

#Estimate daily mean passage for weeks that contain NA

salmon <- salmon \%>\% group_by(wk) \%>\% mutate(wk_days = length(wk))
salmon <- salmon \%>\% group_by(wk_days, year, month, wk) \%>\%
reframe(pd = round(if_else(is.na(pd),mean(pd, na.rm=TRUE),pd))) \%>\%
arrange(wk) \%>\% data.frame()
 salmon

ci_boot_sum(dat = salmon, by = wk, var = pd)
ci_boot_sum(salmon, month, pd)

}
