% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_categorical.R
\name{comma_categorical}
\alias{comma_categorical}
\title{Function to format categorical variables with commas.}
\arguments{
\item{x}{random variable.}
}
\value{
Implement comma format for categorical variables.
}
\description{
Function to format categorical variables with commas.
}
