% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNA.R
\name{countNA}
\alias{countNA}
\title{Count NAs in dataset}
\usage{
countNA(dataframe)
}
\arguments{
\item{dataframe}{NAs by column}
}
\value{
dataframe adding NAs by column
}
\description{
Counts NAs
}
\examples{
library(BD)
dat <- head(airquality, 12)
dat <- as.data.frame(dat)
countNA(dat)
}
