% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNA_total.R
\name{countNA_total}
\alias{countNA_total}
\title{Sum total NAs in dataset}
\usage{
countNA_total(dataframe)
}
\arguments{
\item{dataframe}{A dataset with NAs}
}
\value{
dataframe with total NAs
}
\description{
Counts total NAs in dataframe
}
\examples{
dat <- head(airquality, 12)
dat <- as.data.frame(dat)
countNA_total(dat)

}
