% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_smelt.R
\name{delta_smelt}
\alias{delta_smelt}
\title{shiny app to visualize Delta Smelt}
\usage{
delta_smelt(dat)
}
\arguments{
\item{dat}{Data with multiple columns}
}
\value{
shiny app of Delta Smelt. Type 'delta_smelt(smelt)' to launch shiny app
}
\description{
shiny app to visualize Delta Smelt
}
