% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\docType{data}
\name{ds}
\alias{ds}
\title{Shiny application}
\format{
shiny app to display Delta Smelt Releases:
\describe{
  \item{dat}{Dataset with Delta Smelt}
  ...
}
}
\usage{
ds
}
\description{
Shiny application
}
\keyword{datasets}
