% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{shiny app to visualize any biological dataset}
\usage{
explore(dat)
}
\arguments{
\item{dat}{Data with any number of row and columns}
}
\value{
shiny app to do exploratory data analysis (EDA). Type 'explore(RBsalmon)' or any other dataset in your working directory to launch shiny app
}
\description{
shiny app to visualize any biological dataset
}
