% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents on a state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the state number.}

\item{year}{An integer or a value coercible to an integer representing the year.}
}
\value{
A plot of accidents for the specified state and year. Returns `NULL` if no data is available.
}
\description{
Plots the locations of accidents in a specific state and year.
If there are no accidents or an invalid state number is provided, an appropriate message is displayed.
}
