% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars data by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers or values coercible to integers representing years.}
}
\value{
A tibble summarizing the number of accidents for each month and year.
Columns represent years, and rows represent months.
}
\description{
Summarizes the number of accidents for each year and month.
}
