% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reports.R, R/rb.R
\docType{data}
\name{generate_reports}
\alias{generate_reports}
\title{Generate automated pdf reports of each salmon run}
\format{
Juvenile salmon outmigration:
\describe{
  \item{dataframe}{Requires a data.frame with year and fish runs}
...
}
}
\usage{
generate_reports
}
\arguments{
\item{years}{Can be entered as: years = 2011 or years = 2011:2012.}

\item{var}{Enter the variable to generate report for. eg. var = c('Fall', 'Winter', 'spring')}

\item{input_file}{Rmd file to use to generate reports.}

\item{output_format}{Report with html, word or pdf formats. pdf format is the default.}

\item{output_name}{It will default to the name of the input_file unless the user selects a different name.}
}
\value{
Generate automated pdf reports for individual years and fish runs.
}
\description{
Generate automated pdf reports of each salmon run

Print single or multiple reports
}
\examples{
library(BD)
library(tidyverse)
library(readxl)

my_wd <- system.file("examples", package = "BD")
setwd(my_wd)

# Dataset used to generate report
salmon <- read_excel('RBsalmon.xlsx')
head(salmon)

generate_reports(2012, c("Winter", "Fall"), "individualReport.rmd")

report(years = 2010:2012, var = c("Fall", "Spring"), rmd_file = "individualReport.rmd")
}
\seealso{
\code{\link{report}} to sum multiple years and runs into a single report.

\code{\link{year_var_reports}} Generate automated pdf reports of each salmon run and year.
}
\keyword{datasets}
