% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strips.R
\name{get_strips}
\alias{get_strips}
\title{Color Coding Salinity in San Francisco Bay}
\arguments{
\item{Shapefile}{A shapefile of the desired area. This can be of type .shp or .RDS. Default projection is google mercator (3587).}

\item{StartPoint}{Start point where color coding will start. Make sure to convert coordinates to UTM from lat lon.}

\item{n}{get_strips(n) is the number of strips desired (defaults to 10). Usually the number of desired strips are the number of rows in 'data'}

\item{strip_width}{Width of strips defaults to 1000 meters (1 km). This can be changed in get_strips(strip_width = xxxx)}

\item{strip_length}{Length of strips defaults to 22000 meters (22 km). This can be changed in get_strips(strip_length = xxxx)}

\item{rotate_feature}{Rotate the strips to the desired angle (defaults to 0).}

\item{Mapview/Leaflet}{Use any interactive mapping software or ggplot2 to vizualise shapefile and associated dataset.}

\item{data}{A dataset with 2 rows to overlay on shapefile}

\item{var}{The salinity value per kilometer}
}
\value{
A shapefile associated with a dataframe
}
\description{
Color Coding Salinity in San Francisco Bay
}
\examples{
library(BD)
library(tidyverse)
library(mapview)
library(leaflet)
library(leafem)
library(sf)

head(salinity)
head(sf_bay)
start_UTM <- c(565727.42, 4213365.17)

sf_bay <- sf_bay |> st_transform(26910)
a <- salinity[,2] ;a
 b <- as.numeric(a$freq);b    #convert to numeric to avoid errors
length(b)
the_strips <- get_strips(n = length(b), crs = st_crs(sf_bay)) |> rotate_feature(angle = 105, start_point = start_UTM)
the_strips2 <- cbind(the_strips,salinity)
the_strips2$freq <- as.factor(the_strips2$freq)
the_strips2 <- the_strips2 |> st_intersection(sf_bay)
mapview(the_strips2, zcol = 'freq', layer.name = 'X2 frequency', col.regions= pals::brewer.blues(7))
}
