% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\title{Compute intervals by week}
\arguments{
\item{dataset}{Data with 6 columns}

\item{by}{Compute fish passage by week}

\item{model}{Trap efficiency data}
}
\value{
Confidence intervals by week
}
\description{
Compute intervals by week
}
\examples{
library(BD)
library(tidyverse)
data(fish)
juv <- dplyr::filter(fish, wk < 6)
juv

data(model)
intervals(juv,'wk',model)

}
