% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsmp.R
\name{jsmp}
\alias{jsmp}
\title{shiny app to visualize juvenile salmon runs}
\usage{
jsmp(dat)
}
\arguments{
\item{dat}{Juvenile salmon data with 6 columns.}

\item{jsmp}{Juvenile salmon monitoring program.}
}
\value{
shiny app of salmon runs. Type 'jsmp(RBsalmon)' to launch shiny app
}
\description{
shiny app to visualize juvenile salmon runs
}
