% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a fars filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or a value coercible to an integer representing the year.}
}
\value{
A character string representing the filename.
}
\description{
Generates a filename for a specific year in the FARS dataset. This function permits to create
accident data files by year
}
