% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\docType{data}
\name{model}
\alias{model}
\title{Trap efficiencies}
\format{
A data frame with 119 rows and 2 variables:
\describe{
  \item{PercQ}{Water volume sampled}
  \item{Efficiency}{Number of fish caught divided by total number of fish released}
  ...
}
}
\usage{
model
}
\description{
A dataset containing trap efficiency trials
}
\examples{
library(BD)
dim(model);head(model)
}
\keyword{datasets}
