% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\docType{data}
\name{mydataset}
\alias{mydataset}
\title{Random example}
\format{
A data frame with 2 values:
\describe{
  \item{names}{fictitious names}
  \item{id}{fictitious id's}
  ...
}
}
\usage{
mydataset
}
\description{
A dataset containing names and id
}
\keyword{datasets}
