% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_2_ke.R
\name{na_2_ke}
\alias{na_2_ke}
\title{Convert NA to a number or character}
\arguments{
\item{dat}{data.frame of numerical or mixed formats}

\item{replace}{replace NA in complete data.frame with '.', '-', '0', '?', 'hello'...etc}
}
\value{
data.frame with replaced NA to any number or character
}
\description{
Convert NA to a number or character
}
\examples{

library(BD)
library(dplyr)
data(fish)
juv <- round(data.frame(head(fish)),5)
juv

#Convert NA in data.frame to any desirable value
na_2_ke(juv, '?')
na_2_ke(juv, '_')
na_2_ke(juv, replace = '.')
na_2_ke(juv, 0)
na_2_ke(juv, replace = 'hola')

#To operate only on some columns
juv \%>\%
mutate(xd = ifelse(is.na(xd), '_', xd),
       pd = ifelse(is.na(pd), 'hola', pd))

}
