% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2many.R
\name{one2many}
\alias{one2many}
\title{Deaggregate rows with rownames as integers instead of decimals}
\arguments{
\item{data}{A dataset with multiple rows}

\item{var}{The variable that the user want to split in multiple rows}
}
\value{
An aggregated dataset larger than the original
}
\description{
Deaggregate rows with rownames as integers instead of decimals
}
\examples{
library(tidyverse)
library(BD)
data(RBsalmon)
RBsalmon <- arrange(RBsalmon, SampleDate)
RBsalmon <- RBsalmon[1:20,]
RBsalmon <- data.frame(RBsalmon)
RBsalmon
one2many(RBsalmon, RBsalmon$Count)
}
