% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\docType{data}
\name{onerow2many}
\alias{onerow2many}
\title{Random example}
\format{
A data frame with 2 values:
\describe{
  \item{name}{A fictitious name of people}
  \item{age}{A fictitious age of people}
  ...
}
}
\usage{
onerow2many
}
\description{
A dataset containing names and age
}
\keyword{datasets}
