% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\docType{data}
\name{RBsalmon}
\alias{RBsalmon}
\title{Salmon catch using screw-traps in Red Bluff, CA}
\format{
A data frame with 57,698 rows and 14 columns:
\describe{
  \item{StationCode}{Screwtrap name}
  \item{SampleDate}{Date sample was collected}
  \item{OrganismCode}{Salmon Run code}
  \item{ForkLength}{Fish measurement from snout to tail fork}
  \item{Count}{Number of fish caught}
  \item{race2}{Salmon run, either Spring, Fall, Late-Fall or Winter}

  ...
}
}
\usage{
RBsalmon
}
\description{
A dataset containing raw un-expanded data
}
\examples{
library(BD)
dim(RBsalmon);head(RBsalmon)
}
\keyword{datasets}
