% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_header_gt.R
\name{repeat_header_gt}
\alias{repeat_header_gt}
\title{Long tables headers on every page}
\arguments{
\item{gt}{Table from package gt.}

\item{continuation}{Prints on every page that contains part of the long tables.}
}
\value{
repeat header
}
\description{
Long tables headers on every page
}
