% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate automated pdf report of multiple years and fish runs.}
\arguments{
\item{years}{Can be entered as: years = 2011 or years = 2011:2012.}

\item{var}{Enter the variable to generate report for. eg. var = c('Fall', 'Winter', 'spring')}

\item{input_file}{Rmd file to use to generate reports.}

\item{output_format}{Report with html, word or pdf formats. pdf format is the default.}

\item{output_name}{It will default to the name of the input_file unless the user selects a different name.}
}
\value{
Generate automated pdf reports for individual and multiple years and fish runs as a single report.
}
\description{
Generate automated pdf report of multiple years and fish runs.
}
\examples{
library(BD)
library(tidyverse)
library(readxl)
library(rmarkdown)

my_wd <- system.file("examples", package = "BD")
setwd(my_wd)

# Dataset used to generate report
salmon <- read_excel('RBsalmon.xlsx')
head(salmon)
report(2012, c("Fall"), "individualReport.rmd")

report(years = 2010:2012, var = c("Fall", "Spring"), input_file = "individualReport.rmd")

}
\seealso{
\code{\link{generate_reports}} for other custom pdf report options.

\code{\link{year_var_reports}} Generate automated pdf reports of each salmon run and year.
}
