% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_feature.R
\name{rotate_feature}
\alias{rotate_feature}
\title{Rotating features or strips in San Francisco Bay}
\arguments{
\item{data}{A dataset with 2 rows}

\item{var}{The salinity value per kilometer}
}
\value{
A rotated set of strips on any direction
}
\description{
Rotating features or strips in San Francisco Bay
}
\examples{
library(BD)
library(tidyverse)
library(mapview)
library(leaflet)
library(leafem)

head(salinity);head(sf_bay)
start_UTM <- c(565727.42, 4213365.17)
sf_bay <- sf_bay |> st_transform(26910)
a <- salinity[,2] ;a
 b <- as.numeric(a$freq);b    #convert to numeric to avoid errors
length(b)
the_strips <- get_strips(n = length(b), crs = st_crs(sf_bay)) |> rotate_feature(angle = 105, start_point = start_UTM)
the_strips2 <- cbind(the_strips,salinity)
the_strips2$freq <- as.factor(the_strips2$freq)
the_strips2 <- the_strips2 |> st_intersection(sf_bay)
mapview(the_strips2, zcol = 'freq', layer.name = 'X2 frequency', col.regions= pals::brewer.blues(7))
}
