% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\docType{data}
\name{salinity}
\alias{salinity}
\title{Salinity frequencies}
\format{
A data frame with 45 kilometers:
\describe{
  \item{dist_km}{Kilometers from San Francisco Golden Gate Bridge}
  \item{freq}{Number of salinity values per site}
  ...
}
}
\usage{
salinity
}
\description{
A dataset containing salinity areas
}
\examples{
library(BD)
dim(salinity);head(salinity)
}
\keyword{datasets}
