% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSliderColor.R
\name{setSliderColor}
\alias{setSliderColor}
\title{Enhanced slider}
\usage{
setSliderColor(color, sliderId)
}
\arguments{
\item{years}{Years Sampled}
}
\value{
Range of years selected by user
}
\description{
Enhanced slider
}
