% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\docType{data}
\name{sf_bay}
\alias{sf_bay}
\title{Shape file of the San Francisco Bay}
\format{
Color code based on salinities:
\describe{
  \item{dist_km}{Kilometers from San Francisco Golden Gate Bridge}
  \item{freq}{Number of salinity values per site}
  ...
}
}
\usage{
sf_bay
}
\description{
A shapefile
}
\keyword{datasets}
