% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smelt.R
\name{smelt}
\alias{smelt}
\title{shiny app to visualize Delta Smelt Releases}
\usage{
smelt(dat)
}
\arguments{
\item{dat}{Data with multiple columns}
}
\value{
shiny app of Delta Smelt. Type 'smelt(ds)' to launch the shiny app
}
\description{
shiny app to visualize Delta Smelt Releases
}
