% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_var_reports.R
\name{year_var_reports}
\alias{year_var_reports}
\title{Generate automated pdf reports of each salmon run and year.}
\arguments{
\item{data}{A dataframe with data. Preferably fish data and must have year and salmon race (race2) columns.}

\item{years}{Can be entered as: years = 2011 or years = 2011:2012.}

\item{fish_run}{Enter the variable to generate report for. eg. var = c('Fall', 'Winter', 'spring')}

\item{rmd_file}{Rmd file to use to generate reports.}
}
\value{
Generate automated pdf reports for individual years and fish runs.
}
\description{
Generate automated pdf reports of each salmon run and year.
}
\examples{
library(BD)
library(tidyverse)
library(readxl)
my_wd <- system.file("examples", package = "BD")
setwd(my_wd)
salmon <- read_excel('RBsalmon.xlsx')
salmon <- na.omit(salmon)
salmon$SampleDate <- as.Date(salmon$SampleDate,"\%m/\%d/\%Y")
salmon <- dplyr::filter(salmon,!race == "n/p")
salmon <- dplyr::filter(salmon,ForkLength > 0)
salmon$year <- format(salmon$SampleDate,"\%Y") #year
salmon$week <- format(salmon$SampleDate,"\%V") #week starts with 01 on jan 1..also "\%W" and "\%U" can be used
salmon$month <- format(salmon$SampleDate,"\%B") #month name..abreviated "\%b" or number "\%m"
salmon <- salmon[,c(1,2,5,6,9,11,13,14:16)]
#salmon <- salmon \%>\% dplyr::filter(year >2009 & year < 2012); unique(salmon$year)
head(salmon)
#Line below runs all the years in salmon dataset and all the fish runs (be careful if your dataset contains many years)
year_var_reports(data = salmon, years = unique(data$year), fish_run = unique(data$race2), rmd_file = 'YearAndVariableLoop.rmd')
#Line below prints reports for only they years and runs selected.
year_var_reports(data = salmon, years = c(2009, 2011), fish_run = c('Fall', 'Winter', 'Spring'), rmd_file = 'YearAndVariableLoop.rmd')
#Reports can also be generated without typing the the function arguments but they MUST be in the same order as below.
year_var_reports(salmon, 2010:2011, c('Spring'), 'YearAndVariableLoop.rmd')

}
\seealso{
\code{\link{report}} to sum multiple years and runs into a single report.

\code{\link{generate_reports}} for other custom pdf report options.
}
