% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_boot.r
\name{ci_boot}
\alias{ci_boot}
\title{Compute mean bootstrapped confidence intervals}
\arguments{
\item{dat}{Dataset of juvenile or adult salmon.}

\item{by}{Break dat by week, month or year. These columns must exist in dataset.}

\item{var}{Numeric variable (catch) to compute to estimate mean and confidence intervals.}

\item{conf.int}{Confidence intervals such as 68, 80, 90, 95, 99.}
}
\value{
Mean fish passage and confidence intervals on the mean.
}
\description{
Compute mean bootstrapped confidence intervals
}
\examples{
library(BD)
library(tidyverse)
library(dplyr)
library(Hmisc)
library(conflicted)
conflicts_prefer(dplyr::summarise)
data(fish)
salmon <- dplyr::filter(fish, wk < 6)
salmon

#Estimate daily mean passage for weeks that contain NA

salmon <- salmon \%>\% group_by(wk) \%>\% mutate(wk_days = length(wk))
salmon <- salmon \%>\% group_by(wk_days, year, month, wk) \%>\%
reframe(pd = round(if_else(is.na(pd),mean(pd, na.rm=TRUE),pd))) \%>\%
arrange(wk) \%>\% data.frame()
salmon

ci_boot(dat = salmon, by = wk, var = pd, conf.int = 0.90)
ci_boot(salmon, month, pd)

}
\seealso{
\code{\link{ci_sum}} to compute sum of fish by week, month, and year and traditional confidence intervals.

\code{\link{ci_boot_sum}} to compute summed bootstrapped confidence intervals by week, month, and year.
}
