% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{letter}
\alias{letter}
\title{shiny app to print automated section 7 concurrence letters}
\usage{
letter()
}
\value{
shiny app for documents. Type 'letter()' to launch the shiny interface.
}
\description{
shiny app to print automated section 7 concurrence letters
}
\examples{
letter()
}
