% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\alias{alias}
\title{shiny app to visualize juvenile salmon runs}
\usage{
report(dat)
}
\arguments{
\item{dat}{Data with 6 columns}

\item{SampleDate}{plot daily values}

\item{OrganismCode}{only 'chn' is being analized.}

\item{Forklength}{length of fish from snout to fork}

\item{Count}{daily counts of actual juvenile salmon captured by screw-traps}

\item{race2}{salmon runs of winter, spring, fall, and late-fall}
}
\value{
shiny app of salmon runs. Type 'report(Rbsalmon)' to launch shiny app
}
\description{
shiny app to visualize juvenile salmon runs
}
