% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R, R/report.R
\docType{data}
\name{report}
\alias{report}
\title{Shiny application}
\format{
shiny app to display juvenile salmon outmigration:
\describe{
  \item{dat}{Dataset with juvenile salmon catch}
  ...
}
}
\usage{
report

report(dat)
}
\arguments{
\item{dat}{Data with 6 columns}
}
\value{
shiny app of salmon runs. Type 'report(RBsalmon)' to launch shiny app
}
\description{
Shiny application

shiny app to visualize juvenile salmon runs
}
\keyword{datasets}
