% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate automated pdf reports of each salmon run and year.}
\arguments{
\item{years}{Can be entered as: years = 2011 or years = 2011:2012.}

\item{var}{Enter the variable to generate report for. eg. var = c('Fall', 'Winter', 'spring')}

\item{input_file}{Rmd file to use to generate reports.}

\item{output_format}{Report with html, word or pdf formats. pdf format is the default.}

\item{output_name}{It will default to the name of the input_file unless the user selects a different name.}
}
\value{
Generate automated pdf reports for individual years and fish runs. Can also sum multiple runs for a single year.
}
\description{
Generate automated pdf reports of each salmon run and year.
}
\examples{
library(BD)
library(tidyverse)
library(readxl)

my_wd <- system.file("examples", package = "BD")
setwd(my_wd)

# Dataset used to generate report
salmon <- read_excel('RBsalmon.xlsx')
head(salmon)
report(2010:2012, c("Fall", "Spring"), "individualReport_mult.rmd") #THIS WORKS WELL

}
\seealso{
\code{\link{generate_reports}} to sum multiple years and runs into a single report.

\code{\link{year_var_reports}} for other custom pdf report options.
}
