% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\docType{data}
\name{sf_bay}
\alias{sf_bay}
\title{Shape file of the San Francisco Bay
NOTE TO REMEMBER: This file was created from suisun_rio_vista.RDS and saved as
sf_bay.rda...example: sf_bay <- readRDS('suisun_rio_vista.RDS'); save(sf_bay, file = 'sf_bay.rda')
That's the reason the example for get_strips() starts with sf_bay() without loading it.
A shapefile}
\format{
Color code based on salinities:
\describe{
  \item{dist_km}{Kilometers from San Francisco Golden Gate Bridge}
  \item{freq}{Number of salinity values per site}
  ...
}
}
\usage{
sf_bay
}
\description{
Shape file of the San Francisco Bay
NOTE TO REMEMBER: This file was created from suisun_rio_vista.RDS and saved as
sf_bay.rda...example: sf_bay <- readRDS('suisun_rio_vista.RDS'); save(sf_bay, file = 'sf_bay.rda')
That's the reason the example for get_strips() starts with sf_bay() without loading it.
A shapefile
}
\keyword{datasets}
