% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_smelt.R, R/rb.R
\docType{data}
\name{smelt}
\alias{smelt}
\title{shiny app to visualize Delta Smelt}
\format{
shiny app to display juvenile Delta Smelt:
\describe{
  \item{dat}{Dataset with juvenile Delta Smelt recaptures}
  ...
}
library(BD)
dim(smelt);head(smelt)
}
\usage{
smelt(dat)

smelt
}
\arguments{
\item{dat}{Data with multiple columns}
}
\value{
shiny app of Delta Smelt. Type 'smelt(smelt)' to launch shiny app
}
\description{
shiny app to visualize Delta Smelt

Shiny application
}
\keyword{datasets}
