% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smelt.R
\name{smelt}
\alias{smelt}
\title{shiny app to visualize Delta Smelt Releases}
\usage{
smelt(dat)
}
\arguments{
\item{dat}{Data with multiple columns}
}
\value{
shiny app of Delta Smelt. Type 'smelt(yourdataframe)' to launch the shiny app. Your dataset must have the following fields to work properly:
lat, lon, SampleDate, month, year, numb_fish, Survey, ReleaseEvent, ReleaseMethod, origin, and LifeStage.
}
\description{
shiny app to visualize Delta Smelt Releases
}
\examples{
library(shiny)
library(mapview)
library(sf)
library(ggplot2)
library(dplyr)
library(shinyjs)
library(readxl)
library(shinythemes)
library(shinyWidgets)
library(leaflet)
library(lubridate)
library(BD)

my_wd <- system.file("examples", package = "BD")
setwd(my_wd)
mydat <- read_excel('delta_smelt.xlsx', sheet = "delta_smelt")

#Select only columns of interest
mydat <- mydat[,c(1:2,11:13,19,22)]
head(mydat)

# NOTE: Don't add the columns again if they already exist on your original dataframe
mydat$numb_fish <- rep(1,nrow(mydat)) #add number of fish...each row is one fish
mydat$LifeStage <- with(mydat,ifelse(is.na(LifeStage), 'unkLifeStage', LifeStage))
mydat <- mydat |> rename(lat = LatitudeStart, lon = LongitudeStart)
mydat <- mydat \%>\% dplyr::filter(!is.na(lat) & !is.na(lon))
mydat$SampleDate <- as.Date(mydat$SampleDate,"\%m/\%d/\%Y")
mydat$year <- format(mydat$SampleDate,"\%Y")
mydat$year <- as.numeric(mydat$year)

mydat <- mydat \%>\% mutate(month = floor_date(SampleDate, "month"))

# make mydat sf object:
ds_sf <- st_as_sf(mydat,coords = c(5, 4), remove = FALSE, crs = 4326)
colors <- colorRampPalette(c('yellow', 'red', 'blue', 'green'))

#Display the shiny app
smelt(mydat)

}
